backend/requirements/base.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
redis==5.0.8
celery==5.4.0
django-celery-beat==2.7.0
Pillow==10.4.0
python-decouple==3.8
ocpp==1.0.0
websockets==12.0
channels==4.1.0
daphne==4.1.2
channels-redis==4.2.0
firebase-admin==6.5.0
phonenumbers==8.13.44
django-phonenumber-field==7.3.0
haversine==2.8.1
